add a test internal/external dialog
