+++ core/src/next/cognitiveFunctions.ts
@@ -1,33 +1,7 @@
 import { EnumLike, z } from "zod"
-import { BrainFunction, CortexStep } from "./CortexStep";
+import { CortexStep } from "./CortexStep";
 import { ChatMessageRoleEnum } from "./languageModels";
 import { html } from "common-tags";
-
-const singleResponse = (action:string, description:string) => {
-  return () => {
-    const params = z.object({
-      [action]: z.string().describe(description)
-    })
-
-    return {
-      name: action,
-      description,
-      parameters: params,
-      process: (step: CortexStep<any>, response: z.output<typeof params>) => {
-        return {
-          value: response[action],
-          memories: [{
-            role: ChatMessageRoleEnum.Assistant,
-            content: `${step.entityName} ${action} ${response[action]}`
-          }],
-        }
-      }
-    };
-  }
-}
-
-export const externalDialog = singleResponse
-export const internalMonologue = singleResponse
 
 export const decision = (description:string, choices: EnumLike) => {
   return () => {
@@ -82,11 +56,18 @@ export const brainstorm = (description:string) => {
 export const queryMemory = (query:string) => {
   return () => {
     return {
-      name: "queryMemory",
+      name: "query_memory",
       description: query,
       parameters: z.object({
         answer: z.string().describe(`The answer to: ${query}`)
-      })
+      }),
+      command: html`
+        Do not repeat ${query} and instead use a dialog history.
+        Do not copy sections of the chat history as an answer.
+        Do summarize and thoughtfully answer in sentence and paragraph format.
+        
+        Analyze the chat history step by step and answer the question: ${query}.
+      `
     };
   }
 }
